+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Updates
 
 - Updated the styles on the `Label` component to improve text alignemnt. This update also updates the styles for the label element in the `FeedbackBox`, `NewsletterSignup`, `Select`, `Slider`, and `TextInput` components.
+- Updates the `ComponentWrapper` component so it always renders the `HelperErrorText` component. This resolves aria-live announcement issues when there is invalid text but no helper text.
 
 ## 4.1.5 (March 5, 2026)